requirements.txt
install
flask
flask-sqlalchemy
passlib
cryptography
